% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{percent}
\alias{percent}
\title{Formats a double as percent}
\usage{
percent(x, digits = 0)
}
\arguments{
\item{x}{double value to be formated.}

\item{digits}{integer number of digits after '.'.}
}
\value{
formatted string
}
\description{
Formats a double as percent
}
\examples{

percent(0.8)
}
